feat: Refactor checksum to own submodule
