+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod cache;
+pub mod checksum;
 pub mod notifications;
 mod persistence;
 pub mod query;
