+++ src/utils.rs
@@ -1,16 +1,13 @@
 use crate::structs::Resource;
 use anyhow::{anyhow, Result};
-use aruna_rust_api::api::storage::models::v2::{generic_resource, User};
 use aruna_rust_api::api::{
     notification::services::v2::Resource as ApiResource,
     storage::models::v2::{InternalRelation, RelationDirection, ResourceVariant},
 };
-use base64::{engine::general_purpose, Engine};
 use diesel_ulid::DieselUlid;
 use std::str::FromStr;
 use tonic::metadata::AsciiMetadataKey;
 use tonic::metadata::AsciiMetadataValue;
-use xxhash_rust::xxh3::xxh3_128;
 
 pub trait GetRef {
     fn get_ref(&self) -> Option<(DieselUlid, Resource)>;
@@ -106,36 +103,4 @@ impl tonic::service::Interceptor for ClientInterceptor {
 
         Ok(mut_req)
     }
-}
-
-pub fn checksum_resource(gen_res: generic_resource::Resource) -> Result<String> {
-    match gen_res {
-        generic_resource::Resource::Project(mut proj) => {
-            proj.stats = None;
-            Ok(general_purpose::STANDARD_NO_PAD
-                .encode(xxh3_128(&bincode::serialize(&proj)?).to_be_bytes())
-                .to_string())
-        }
-        generic_resource::Resource::Collection(mut col) => {
-            col.stats = None;
-            Ok(general_purpose::STANDARD_NO_PAD
-                .encode(xxh3_128(&bincode::serialize(&col)?).to_be_bytes())
-                .to_string())
-        }
-        generic_resource::Resource::Dataset(mut ds) => {
-            ds.stats = None;
-            Ok(general_purpose::STANDARD_NO_PAD
-                .encode(xxh3_128(&bincode::serialize(&ds)?).to_be_bytes())
-                .to_string())
-        }
-        generic_resource::Resource::Object(obj) => Ok(general_purpose::STANDARD_NO_PAD
-            .encode(xxh3_128(&bincode::serialize(&obj)?).to_be_bytes())
-            .to_string()),
-    }
-}
-
-pub fn checksum_user(user: &User) -> Result<String> {
-    Ok(general_purpose::STANDARD_NO_PAD
-        .encode(xxh3_128(&bincode::serialize(user)?).to_be_bytes())
-        .to_string())
 }
